Fix stale surface holder.
The problem was, when we are back to onStart after onPause (say, after
turning off the screen and back on again), the camera was correctly
released and re-acquired, but its surface holder reference gone! So
we have to remember the state of the surface holder within its creation
and destruction (which spans longer than onPause -> onStart), and set
the newly acquired camera's surface holder.